test: move template-provider mempool test
Move test_create_mempool_transaction into template_provider_integration.

Behavior changes:
- The test name and assertions stay the same.
