+++ integration-tests/lib/template_provider.rs
@@ -506,18 +506,3 @@ impl Drop for TemplateProvider {
         // bitcoin-node is managed by corepc-node::Node and will be cleaned up automatically
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::{DifficultyLevel, TemplateProvider};
-    use crate::utils::get_available_address;
-
-    #[tokio::test]
-    async fn test_create_mempool_transaction() {
-        let address = get_available_address();
-        let port = address.port();
-        let tp = TemplateProvider::start(port, 1, DifficultyLevel::Low);
-        assert!(tp.fund_wallet().is_ok());
-        assert!(tp.create_mempool_transaction().is_ok());
-    }
-}
